requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
black==24.8.0
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
h11==0.14.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
iniconfig==2.0.0
msgpack==1.0.8
mypy-extensions==1.0.0
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==1.7.1
pyOpenSSL==24.2.1
PySocks==1.7.1
pytest==8.3.2
pytest-asyncio==0.24.0
redis==5.0.8
selenium==4.24.0
service-identity==24.1.0
setuptools==74.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.0.3
